A few more optimization of the callback / buffer handling (#269)
* Mark some struct fields as const

This allow the compiler to skip the GC root for the fields values.

* Cache and reuse array wrapper of C buffers

This is based on the observation that the nlopt library frequently
uses the same buffer when communicating with the callback.
The cached arrays may contain dangling references outside of the callback
and should never be accessed by the user.

* Add argument type signature to internal function

Co-authored-by: Oscar Dowson <odow@users.noreply.github.com>

---------

Co-authored-by: Oscar Dowson <odow@users.noreply.github.com>